Refactor: MapStruct mapper, page assembler, gateway rename

+++ src/main/java/com/patken/api/url_shortener/entity/UrlEntity.java
@@ -21,7 +21,7 @@ public class UrlEntity extends CommonEntity implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "url_id")
-    private Integer urlId;
+    private Long urlId;
 
     @Column(name = "original_url", nullable = false, length = 2048)
     private String originalUrl;
